--- Assets/Scripts/Camera/FinalCameraActivator.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: d06647921ab94d0e9df4633cc2c98b87
-timeCreated: 1679242237
