+++ Assets/Scripts/Camera/StaticCameraActivator.cs.meta
