+++ Assets/Scripts/InputReader.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using Player;
 using UnityEngine;
