+++ ProjectSettings/TagManager.asset
@@ -5,6 +5,8 @@ TagManager:
   serializedVersion: 2
   tags:
   - Ground
+  - StartCamera
+  - FinalCamera
   layers:
   - Default
   - TransparentFX
